+++ vk.application/console/inc/Console.h
@@ -13,6 +13,7 @@
 #include "Mutex.h"
 #include "Regex.h"
 #include "Cmd.h"
+#include "List.h"
 
 
 ///Console
