+++ vk.kernel/binutils/inc/LibraryTool.h
@@ -8,8 +8,8 @@
 #define __LIBRARY_TOOL_H__
 
 #include "Defines.h"
-#include "Templates.h"
 #include "ElfLoader.h"
+#include "List.h"
 
 
 /// @brief LibraryTool
