+++ vk.kernel/core/if/Cast.h
@@ -1,0 +1,40 @@
+//###########################################################################
+// Cast.h
+// Declarations of the cast
+//
+// $Copyright: Copyright (C) village
+//###########################################################################
+#ifndef __CAST_H__
+#define __CAST_H__
+
+#include "stdint.h"
+#include "stddef.h"
+#include "stdlib.h"
+
+
+/// @brief Get method address
+/// @tparam dst_type 
+/// @tparam src_type 
+/// @param src 
+/// @return addr 
+template<typename dst_type, typename src_type>
+dst_type pointer_cast(src_type src)
+{
+	return *static_cast<dst_type*>(static_cast<void*>(&src));
+}
+
+
+/// @brief Get method address
+/// @tparam dst_type 
+/// @tparam src_type 
+/// @param src 
+/// @return addr 
+template<typename dst_type, typename src_type>
+dst_type union_cast(src_type src)
+{
+	union{ src_type s; dst_type d; } u;
+	u.s = src;
+	return u.d;
+}
+
+#endif //!__CAST_H__
