+++ vk.kernel/core/inc/Memory.h
@@ -9,6 +9,7 @@
 
 #include "Module.h"
 
+
 /// @brief Memory
 class Memory : public Module
 {
