+++ vk.kernel/core/inc/Thread.h
@@ -9,7 +9,8 @@
 
 #include "Module.h"
 #include "Registers.h"
-#include "Templates.h"
+#include "List.h"
+
 
 #ifndef TASK_STACK
 #define TASK_STACK      1024
