+++ vk.kernel/core/src/Environment.cpp
@@ -5,9 +5,9 @@
 // $Copyright: Copyright (C) village
 //###########################################################################
 #include "Environment.h"
-#include "Templates.h"
 #include "Kernel.h"
 #include "string.h"
+#include "List.h"
 
 
 /// @brief ConcreteEnviroment
