+++ vk.kernel/core/src/Thread.cpp
@@ -10,6 +10,8 @@
 #include "Memory.h"
 #include "System.h"
 #include "Scheduler.h"
+#include "Cast.h"
+#include "List.h"
 
 
 /// @brief ConcreteThread
