+++ vk.kernel/subsys/inc/Loader.h
@@ -8,8 +8,8 @@
 #define __LOADER_H__
 
 #include "Module.h"
-#include "Templates.h"
 #include "ElfExecutor.h"
+#include "List.h"
 
 
 /// @brief Loader
