+++ vk.kernel/subsys/src/Input.cpp
@@ -6,7 +6,8 @@
 //###########################################################################
 #include "Input.h"
 #include "Kernel.h"
-#include "Templates.h"
+#include "List.h"
+#include "Cast.h"
 
 
 /// @brief ConcreteInput
