+++ vk.kernel/subsys/src/Loader.cpp
@@ -7,10 +7,10 @@
 #include "Loader.h"
 #include "Kernel.h"
 #include "rcParser.h"
-#include "Templates.h"
 #include "LibraryTool.h"
 #include "ModuleTool.h"
 #include "ElfExecutor.h"
+#include "List.h"
 
 
 /// @brief ConcreteLoader
